requirements.txt
attrs==19.3.0
Automat==0.8.0
autopep8==1.5.4
backcall==0.2.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Admin==1.5.7
Flask-SQLAlchemy==2.4.4
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
openpyxl==3.0.6
pandas==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.6.0
Pygments==2.7.4
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
pyzmq==22.0.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.23
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
xlrd==2.0.1
zipp==1.0.0
zope.interface==4.7.1
